refactor(Token): 토큰 관련 로직들을 token 패키지로 이동
