+++ src/main/java/shopping/token/application/LoginRequest.java
@@ -1,4 +1,4 @@
-package shopping.auth.application;
+package shopping.token.application;
 
 public record LoginRequest(String email, String password) {
 }
