+++ src/main/java/shopping/token/presentation/TokenController.java
@@ -1,24 +1,24 @@
-package shopping.auth.presentation;
+package shopping.token.presentation;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import shopping.auth.application.AuthenticationService;
-import shopping.auth.application.LoginRequest;
+import shopping.token.application.TokenService;
+import shopping.token.application.LoginRequest;
 import shopping.token.domain.Token;
 
 @RestController
-public class AuthenticationController {
-    private final AuthenticationService authenticationService;
+public class TokenController {
+    private final TokenService tokenService;
 
-    public AuthenticationController(AuthenticationService authenticationService) {
-        this.authenticationService = authenticationService;
+    public TokenController(TokenService tokenService) {
+        this.tokenService = tokenService;
     }
 
     @PostMapping("/login")
     public ResponseEntity<TokenResponse> login(@RequestBody LoginRequest loginRequest) {
-        Token token = authenticationService.authenticate(loginRequest);
+        Token token = tokenService.generate(loginRequest);
         return ResponseEntity.ok(TokenResponse.from(token));
     }
 }
