+++ src/main/java/shopping/token/presentation/TokenResponse.java
@@ -1,4 +1,4 @@
-package shopping.auth.presentation;
+package shopping.token.presentation;
 
 import shopping.token.domain.Token;
